% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consr.R
\name{consensus}
\alias{consensus}
\title{Consensus for an item}
\usage{
consensus(item, range = NULL, consensus.only = FALSE, round = 2)
}
\arguments{
\item{item}{The item that should be analyzed.}

\item{range}{The range of the item indicating the valid values. Default is \code{NULL}.}

\item{consensus.only}{Logical, if only the consensus should be reported. Default is to \code{FALSE}.}

\item{round}{The number of digits to round the result to. Default is to 2.}
}
\value{
Returns an data frame with the consensus and additional information for each item. See details for further explanation.
}
\description{
Compute the consensus for an item, based on Tastle & Wiermann, (2006).
}
\details{
\describe{
 \item{n}{The number of valid answers on that item}
 \item{min}{The minimum answer on that item}
 \item{max}{The maximum answer on that item}
 \item{consensus}{The consensus}
 \item{dissensus}{The dissensus}
 \item{n_missing}{The absolute number of missing values on that item, defined as \code{NA}}
 \item{perc_missing}{The relative number of missing values on that item, defined as \code{NA}}
}
}
\examples{
item <- c(rep(1, 19), rep(2, 8), rep(3, 26), rep(4, 29),
          rep(5, 5), rep(NA, 5), rep(-2, 4), rep(-3, 4))
consensus(item, range = c(1, 5), round = 4)
## or the original example from the article:
item <- c(rep(1, 19), rep(2, 16), rep(3, 26), rep(4, 29), rep(5, 10))
consensus(item, range = c(1, 5), round = 3)
}

