% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consr.R
\name{freq_item}
\alias{freq_item}
\title{Get frequencies in an item and plot them}
\usage{
freq_item(item, range = NULL, plot = FALSE, ...)
}
\arguments{
\item{item}{The item to be analysed. Must be a vector.}

\item{range}{The range of valid values.}

\item{plot}{logical, if a plot should be plotted. Defaultis to \code{FALSE}.}

\item{...}{Further arguments passed on to \code{\link[graphics]{barplot}}.}
}
\value{
Returns a frequency table of the entries in that item. Optionally, also a barplot of these items.
}
\description{
Get frequencies in an item and plot them
}
\examples{
item <- c(rep(1, 19), rep(2, 8), rep(3, 26), rep(4, 29),
          rep(5, 5), rep(NA, 5), rep(-2, 4), rep(-3, 4))
freq_item(item, range = c(1,5), plot = TRUE, col = "#00376c")
}

