% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consr.R
\name{freq_overview}
\alias{freq_overview}
\title{Get frequencies from an item}
\usage{
freq_overview(item, return.df = FALSE)
}
\arguments{
\item{item}{The item that will be analysed}

\item{return.df}{Logical, should a data frame with the overview be returned? Default is fot \code{FALSE}.}
}
\value{
An overview for the frequencies of each entry in the item.
}
\description{
The return value contains two enries for each category used in the data. One that ends with \code{abs} shows the absolute frequencies, the other shows the relative frequencies, ending on \code{perc}. \code{min} before an entry means that this value was negative. Example: \code{C_min2_perc} with a value of 5 means that the -2 was present 5% on that item.
}
\examples{
item <- c(rep(1, 19), rep(2, 8), rep(3, 26), rep(4, 29),
          rep(5, 5), rep(NA, 5), rep(-2, 4), rep(-3, 4))
freq_overview(item, return.df = TRUE)
}

