% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consr.R
\name{integer_subset_df}
\alias{integer_subset_df}
\title{Subset integers from data frame}
\usage{
integer_subset_df(data)
}
\arguments{
\item{data}{a data frame or matrix.}
}
\value{
The same data frame as \code{data}, but the non-integers are dropped.
}
\description{
For computing the consensus, only integers should be included in the data frame.
}
\examples{
a <- sample(1:20, 10)
b <- rnorm(10)
c <- sample(-2:2, 10, replace = TRUE)
integer_subset_df(data.frame(a, b, c))
}
\seealso{
If you need to recode data into integers, see for example \code{\link[car]{recode}}.
}

