% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consr.R
\name{prep_item}
\alias{prep_item}
\title{Prepare an item for consensus analysis}
\usage{
prep_item(item, range = NULL)
}
\arguments{
\item{item}{The item to be recoded}

\item{range}{The range of the Likert-scale for that item. Must be a vector of at least two characters.}
}
\value{
The item itself, where all entries outside of \code{range} are recoded as \code{NA}.
}
\description{
Prepare an item for consensus analysis
}
\examples{
item <- c(rep(1, 19), rep(2, 8), rep(3, 26), rep(4, 29),
          rep(5, 5), rep(NA, 5), rep(-2, 4), rep(-3, 4))
prep_item(item, range = c(1, 5)
}

