% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consr.R
\name{rbind_map}
\alias{rbind_map}
\title{rbind data frames from lists}
\usage{
rbind_map(list)
}
\arguments{
\item{list}{A list where each entry is a data frame with the same columns}
}
\value{
A data frame.
}
\description{
rbind data frames from lists
}

